% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ci}
\alias{ci}
\title{Calculate Credible Intervals (wide and narrow).}
\usage{
ci(D, thick_ci = c(0.05, 0.95), thin_ci = c(0.025, 0.975))
}
\arguments{
\item{D}{Data frame whith the simulations.}

\item{thick_ci}{Vector of length 2 with the quantiles of the thick band for the credible interval}

\item{thin_ci}{Vector of length 2 with the quantiles of the thin band for the credible interval}
}
\value{
A data frame tibble with the Parameter names and 5 variables with the limits of the credibal intervals (thin and thick), ready to be used to produce caterpillar plots.
}
\description{
Generate a data frame with the limits of two credible intervals. Function used by \code{\link{ggs_caterpillar}}. "low" and "high" refer to the wide interval, whereas "Low" and "High" refer to the narrow interval. "median" is self-explanatory and is used to draw a dot in caterpillar plots. The data frame generated is of wide format, suitable for ggplot2::geom_segment().
}
\examples{
data(linear)
ci(ggs(s))
}
