% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggs.R
\name{custom.sort}
\alias{custom.sort}
\title{Auxiliary function that sorts Parameter names taking into account numeric values}
\usage{
custom.sort(x)
}
\arguments{
\item{x}{a character vector to which we want to sort elements}
}
\value{
X a character vector sorted with family parametrs first and then numeric values
}
\description{
Auxiliary function that sorts Parameter names taking into account numeric values
}

