% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{get_family}
\alias{get_family}
\title{Subset a ggs object to get only the parameters with a given regular expression.}
\usage{
get_family(D, family = NA)
}
\arguments{
\item{D}{Data frame with the data arranged and ready to be used by the rest of the ggmcmc functions. The dataframe has four columns, namely: Iteration, Parameter, value and Chain, and six attributes: nChains, nParameters, nIterations, nBurnin, nThin and description.}

\item{family}{Name of the family of parameters to plot, as given by a character vector or a regular expression. A family of parameters is considered to be any group of parameters with the same name but different numerical value between square brackets (as beta[1], beta[2], etc).}
}
\value{
D Data frame that is a subset of the given D dataset.
}
\description{
Internal function used by the graphical functions to get only some of the parameters that follow a given regular expression.
}

