\name{gl.unq}
\alias{gl.unq}
\title{Generate a factor with unequal number of repetitions}
\usage{
  gl.unq(n, k, labels = 1:n)
}
\arguments{
  \item{n}{number of levels}

  \item{k}{number of repetitions}

  \item{labels}{optional vector of labels}
}
\value{
  A factor
}
\description{
  Generate a factor with levels of unequal length
}

