% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{gl_unq}
\alias{gl_unq}
\title{Generate a factor with unequal number of repetitions.}
\usage{
gl_unq(n, k, labels = 1:n)
}
\arguments{
\item{n}{number of levels}

\item{k}{number of repetitions}

\item{labels}{optional vector of labels}
}
\value{
A factor
}
\description{
Generate a factor with levels of unequal length.
}
\details{
Internal function to generate a factor with levels of unequal length, used by \code{\link{ggs_histogram}}.
}

