% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{jgmcmc}
\alias{jgmcmc}
\alias{jgmcmc-package}
\alias{jgmcmc-package}
\title{Graphical tools for analyzing Markov Chain Monte Carlo simulations from Bayesian inference.}
\description{
ggmcmc is a tool for assessing and diagnosing convergence of
Markov Chain Monte Carlo simulations, as well as for graphically display
results from full MCMC analysis. The package also facilitates the graphical
interpretation of models by providing flexible functions to plot the
results against observed variables.
}
\references{
\url{http://xavier-fim.net/packages/ggmcmc}.
}
