% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggs_rocplot.R
\name{roc_calc}
\alias{roc_calc}
\title{Calculate the ROC curve for a set of observed outcomes and predicted probabilities}
\usage{
roc_calc(R)
}
\arguments{
\item{R}{data frame with the 'value' (predicted probability) and the observed 'Outcome'.}
}
\value{
A data frame with the Sensitivity and the Specificity.
}
\description{
Internal function used by \code{\link{ggs_autocorrelation}}.
}
\references{
Fernández-i-Marín, Xavier (2016) ggmcmc: Analysis of MCMC Samples and Bayesian Inference. Journal of Statistical Software, 70(9), 1-20. doi:10.18637/jss.v070.i09
}
