\docType{data}
\name{S}
\alias{S}
\title{Simulations of the parameters of a simple linear regression with fake data.}
\format{A coda object with 2 chains of 1000 iterations each one for 3 parameters}
\description{
  A coda object containing simulated values from the
  posterior distribution of the intercept, slope and error
  of a simple linear regression with fake data (y = beta[0]
  + beta[1] * X + sigma) . The purpose of the dataset is
  only to show the possibilities of the ggmcmc package.
}
\keyword{datasets}

