% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s.y.rep}
\alias{s.y.rep}
\title{Simulations of the posterior predictive distribution of a simple linear regression with fake data.}
\format{
A coda object containing posterior distributions of the posterior predictive distribution of a linear regression with fake data.
}
\usage{
data(s.y.rep)
}
\description{
A coda object containing simulated values from the posterior predictive distribution of the outcome of a linear regression with fake data (y ~ N(mu, sigma); mu = beta[1] + beta[2] * X; y.rep ~ N(mu, sigma); where y.rep is a replicated outcome, originally missing data). The purpose of the dataset is only to show the possibilities of the ggmcmc package.
}
\keyword{datasets}
