\name{sde0f}
\alias{sde0f}
\title{Spectral Density Estimate at Zero Frequency}
\usage{
  sde0f(x)
}
\arguments{
  \item{x}{A time series}
}
\value{
  A vector with the spectral density estimate at zero
  frequency
}
\description{
  Compute the Spectral Density Estimate at Zero Frequency
  for a given chain.
}

