% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y}
\alias{y}
\title{Values for the observed outcome of a simple linear regression with fake data.}
\format{
A numeric vector containing the observed values of the outcome in the linear regression with fake data.
}
\usage{
y
}
\description{
A numeric vector containing the observed values of the outcome of a linear regression with fake data (y = beta[1] + beta[2] + X + sigma). The purpose of the dataset is only to show the possibilities of the ggmcmc package.
}
\keyword{datasets}
