% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggs.R
\name{ggs_chain}
\alias{ggs_chain}
\title{Auxiliary function that extracts information from a single chain.}
\usage{
ggs_chain(s)
}
\arguments{
\item{s}{a single chain to convert into a data frame}
}
\value{
D data frame with the chain arranged
}
\description{
Auxiliary function that extracts information from a single chain.
}

