% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggs_running.R
\name{ggs_running}
\alias{ggs_running}
\title{Running means of the chains}
\usage{
ggs_running(D, family = NA, original_burnin = TRUE,
  original_thin = TRUE, greek = FALSE)
}
\arguments{
\item{D}{Data frame whith the simulations.}

\item{family}{Name of the family of parameters to plot, as given by a character vector or a regular expression. A family of parameters is considered to be any group of parameters with the same name but different numerical value between square brackets (as beta[1], beta[2], etc).}

\item{original_burnin}{Logical. When TRUE (the default), start the iteration counter in the x-axis at the end of the burnin period.}

\item{original_thin}{Logical. When TRUE (the default), take into account the thinning interval in the x-axis.}

\item{greek}{Logical value indicating whether parameter labels have to be parsed to get Greek letters. Defaults to false.}
}
\value{
A \code{ggplot} object.
}
\description{
Running means of the chains.
}
\examples{
data(linear)
ggs_running(ggs(s))
}
\references{
Fernández-i-Marín, Xavier (2016) ggmcmc: Analysis of MCMC Samples and Bayesian Inference. Journal of Statistical Software, 70(9), 1-20. doi:10.18637/jss.v070.i09
}
