% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggs_rocplot.R
\name{roc_calc}
\alias{roc_calc}
\title{Calculate the ROC curve for a set of observed outcomes and predicted probabilities}
\usage{
roc_calc(R)
}
\arguments{
\item{R}{data frame with the 'value' (predicted probability) and the observed 'Outcome'.}
}
\value{
A data frame with the Sensitivity and the Specificity.
}
\description{
Internal function used by \code{\link{ggs_autocorrelation}}.
}

