% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s}
\alias{s}
\title{Simulations of the parameters of a simple linear regression with fake data.}
\format{
A coda object containing posterior distributions of the intercept, slope and residual of a linear regression with fake data.
}
\usage{
s
}
\description{
A coda object containing simulated values from the posterior distribution of the intercept, slope and residual of a linear regression with fake data (y = beta[1] + beta[2] * X + sigma). The purpose of the dataset is only to show the possibilities of the ggmcmc package.
}
\keyword{datasets}
