% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sde0f}
\alias{sde0f}
\title{Spectral Density Estimate at Zero Frequency.}
\usage{
sde0f(x)
}
\arguments{
\item{x}{A time series}
}
\value{
A vector with the spectral density estimate at zero frequency
}
\description{
Compute the Spectral Density Estimate at Zero Frequency for a given chain.
}
\details{
Internal function to compute the Spectral Density Estimate at Zero Frequency for a given chain used by \code{\link{ggs_geweke}}.
}
